Move all user authentication related code to an Authenticator class.
