+++ Controllers/AuthController.cs
@@ -10,6 +10,7 @@ using System.Security.Claims;
 using Entities;
 using Microsoft.IdentityModel.Tokens;
 using Viscon_ProjectC_Groep4.Dto;
+using Services;
 
 namespace Viscon_ProjectC_Groep4.Controllers {
     [ApiController]
@@ -17,11 +18,16 @@ namespace Viscon_ProjectC_Groep4.Controllers {
     public class AuthController : ControllerBase {
         private readonly IConfiguration _configuration;
         private readonly ILogger<AuthController> _logger;
+        private readonly Authenticator _authenticator;
         private readonly IServiceProvider _services;
 
-        public AuthController(IConfiguration configuration, ILogger<AuthController> logger, IServiceProvider services) {
+        public AuthController(
+            IConfiguration configuration, ILogger<AuthController> logger,
+            Authenticator authenticator, IServiceProvider services
+        ) {
             _configuration = configuration;
             _logger = logger;
+            _authenticator = authenticator;
             _services = services;
         }
 
@@ -38,14 +44,17 @@ namespace Viscon_ProjectC_Groep4.Controllers {
                     return BadRequest("User not found");
                 }
 
-                if (!VerifyPassword(data.Password, user.Usr_Password, user.Usr_PasswSalt)) {
+                if (!_authenticator.VerifyPassword(
+                    data.Password, user.Usr_Password, user.Usr_PasswSalt
+                )) {
                     _logger.LogError("Wrong Password");
                     return BadRequest("Wrong password");
                 }
                 _logger.LogInformation("User and Passw correct", data.Password);
 
-                var token = CreateToken(user);
-                    
+                _logger.LogInformation("Creating user token");
+                var token = _authenticator.CreateToken(user);
+
                 return Ok(token);
             }
             catch (Exception ex) {
@@ -59,7 +68,9 @@ namespace Viscon_ProjectC_Groep4.Controllers {
             try {
                 _logger.LogInformation(data.FirstName + " " + data.LastName + " is trying to be created");
 
-                CreatePassHash(data.Password, out byte[] passwordHash, out byte[] passwordSalt);
+                _authenticator.CreatePassHash(
+                    data.Password, out byte[] passwordHash, out byte[] passwordSalt
+                );
 
                 _logger.LogInformation(data.Password, passwordHash, passwordSalt);
 
@@ -95,73 +106,6 @@ namespace Viscon_ProjectC_Groep4.Controllers {
             }
             catch (Exception ex) {
                 return BadRequest(ex.Message);
-            }
-        }
-        
-        private string CreateToken(User user) {
-
-            _logger.LogInformation($"Creating Token for {user.Usr_Email}");
-
-            List<Claim> claims = new List<Claim> {
-                new Claim(ClaimTypes.NameIdentifier, user.Usr_Id.ToString()),
-                new Claim(ClaimTypes.Name, user.Usr_FirstName),
-                new Claim(ClaimTypes.Name, user.Usr_LastName),
-                new Claim(ClaimTypes.Email, user.Usr_Email),
-                new Claim(ClaimTypes.MobilePhone, user.Usr_PhoneNumber.ToString()),
-                new Claim(ClaimTypes.Role, user.Usr_Role.ToString()),
-                new Claim("DepartmentId", user.Usr_DepId.ToString()),
-                new Claim("CompanyId", user.Usr_CompId.ToString()),
-            };
-
-            var key = new SymmetricSecurityKey(System.Text.Encoding.UTF8
-                .GetBytes("aSlOdjn0gJKhZBerB9oPDiCxQ7h5aRmYyNX94ytbxC7HJmq109VREMVAJW+dpKpSbZZryJPbHgX7jpy6bLvy0Iw="));
-
-            var creds = new SigningCredentials(key, SecurityAlgorithms.HmacSha512Signature);
-            
-            try {
-                var token = new JwtSecurityToken(
-                    claims: claims,
-                    expires: DateTime.Now.AddDays(1),
-                    signingCredentials: creds
-                );
-
-                var tokenHandler = new JwtSecurityTokenHandler();
-                string tokenString = tokenHandler.WriteToken(token);
-                _logger.LogInformation($"Token created successfully. {token}");
-                return tokenString;
-            } catch (Exception ex) {
-                _logger.LogError($"Error creating token: {ex.Message}");
-                return null!;
-            }
-            
-        }
-
-        public static bool VerifyToken(string token, out int Id) {
-            Id = 0;
-
-            var pre = new JwtSecurityTokenHandler();
-            var jwt = pre.ReadJwtToken(token);
-            
-            if (jwt == null || jwt.ValidFrom > DateTime.UtcNow || jwt.ValidTo < DateTime.UtcNow) {
-                return false;
-            }
-
-            var claim = jwt.Claims.First(claim => claim.Type == ClaimTypes.NameIdentifier).Value;
-            Id = Int32.Parse(claim);
-            return true;
-        }
-
-        private void CreatePassHash(string password, out byte[] passwordHash, out byte[] passwordSalt) {
-            using (var hmac = new HMACSHA512()) {
-                passwordSalt = hmac.Key;
-                passwordHash = hmac.ComputeHash(System.Text.Encoding.UTF8.GetBytes(password));
-            }
-        }
-
-        private bool VerifyPassword(string password, byte[] passwordHash, byte[] passwordSalt) {
-            using (var hmac = new HMACSHA512(passwordSalt)) {
-                var CalcHash = hmac.ComputeHash(System.Text.Encoding.UTF8.GetBytes(password));
-                return CalcHash.SequenceEqual(passwordHash);
             }
         }
     }
