+++ Controllers/TicketController.cs
@@ -1,6 +1,6 @@
 using Microsoft.AspNetCore.Mvc;
-using static Viscon_ProjectC_Groep4.Controllers.AuthController;
 using Entities; // Make sure this namespace exists
+using Services;
 using Viscon_ProjectC_Groep4.Dto;
 
 namespace Viscon_ProjectC_Groep4.Controllers
@@ -10,20 +10,24 @@ namespace Viscon_ProjectC_Groep4.Controllers
     public class TicketController : ControllerBase
     {
         private readonly ILogger<TicketController> _logger;
+        private readonly Authenticator _authenticator;
         private readonly IServiceProvider _services;
 
-        public TicketController(ILogger<TicketController> logger, IServiceProvider services) {
+        public TicketController(
+            ILogger<TicketController> logger, Authenticator authenticator,
+            IServiceProvider services
+        ) {
             _logger = logger;
+            _authenticator = authenticator;
             _services = services;
         }
-            
             
         [HttpPost("createticket")]
         public async Task<ActionResult<Ticket>> CreateTicket([FromBody] MachineDataDto data) {
             _logger.LogInformation("API Fetched");
             await using var context = _services.GetService<ApplicationDbContext>();
             try {
-                if (VerifyToken(data.Jtw, out int Id)) {
+                if (_authenticator.VerifyToken(data.Jtw, out int Id)) {
                     _logger.LogInformation("Token Correct");
                     var ticket = new Ticket();
                     ticket.Tick_MachId = context.Machines.Where(m => m.Mach_Name == data.machine).Select(m => m.Mach_Id).FirstOrDefault();
